Add card fields (WIP)
