--- modules/ppcp-card-fields/.babelrc
@@ -1,14 +1,0 @@
-{
-	"presets": [
-		[
-			"@babel/preset-env",
-			{
-				"useBuiltIns": "usage",
-				"corejs": "3.25.0"
-			}
-		],
-		[
-			"@babel/preset-react"
-		]
-	]
-}
