--- modules/ppcp-card-fields/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-assets/js
-assets/css
