+++ package.json
@@ -10,7 +10,6 @@
     "install:modules:ppcp-applepay": "cd modules/ppcp-applepay && yarn install",
     "install:modules:ppcp-blocks": "cd modules/ppcp-blocks && yarn install",
     "install:modules:ppcp-button": "cd modules/ppcp-button && yarn install",
-    "install:modules:ppcp-card-fields": "cd modules/ppcp-card-fields && yarn install",
     "install:modules:ppcp-googlepay": "cd modules/ppcp-googlepay && yarn install",
     "install:modules:ppcp-wc-gateway": "cd modules/ppcp-wc-gateway && yarn install",
     "install:modules:ppcp-webhooks": "cd modules/ppcp-webhooks && yarn install",
@@ -22,7 +21,6 @@
     "build:modules:ppcp-applepay": "cd modules/ppcp-applepay && yarn run build",
     "build:modules:ppcp-blocks": "cd modules/ppcp-blocks && yarn run build",
     "build:modules:ppcp-button": "cd modules/ppcp-button && yarn run build",
-    "build:modules:ppcp-card-fields": "cd modules/ppcp-card-fields && yarn run build",
     "build:modules:ppcp-googlepay": "cd modules/ppcp-googlepay && yarn run build",
     "build:modules:ppcp-wc-gateway": "cd modules/ppcp-wc-gateway && yarn run build",
     "build:modules:ppcp-webhooks": "cd modules/ppcp-webhooks && yarn run build",
@@ -35,7 +33,6 @@
     "watch:modules:ppcp-applepay": "cd modules/ppcp-applepay && yarn run watch",
     "watch:modules:ppcp-blocks": "cd modules/ppcp-blocks && yarn run watch",
     "watch:modules:ppcp-button": "cd modules/ppcp-button && yarn run watch",
-    "watch:modules:ppcp-card-fields": "cd modules/ppcp-card-fields && yarn run watch",
     "watch:modules:ppcp-googlepay": "cd modules/ppcp-googlepay && yarn run watch",
     "watch:modules:ppcp-wc-gateway": "cd modules/ppcp-wc-gateway && yarn run watch",
     "watch:modules:ppcp-webhooks": "cd modules/ppcp-webhooks && yarn run watch",
